+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added `Parameter` type for hyper parameters
 * Redefined all kernels as collections of `Parameter`s internally
+* Removed ARD type
+* Removed kernel approximation
 
 # MLKernels 0.1.1 Release Notes